// packages/sdk/gen/tfplugin6.ts
export interface DynamicValue {
  msgpack: Uint8Array;
  json: Uint8Array;
}

export interface Diagnostic {
  severity: Diagnostic_Severity;
  summary: string;
  detail: string;
  attribute: AttributePath | undefined;
}

export enum Diagnostic_Severity {
  INVALID = 0,
  ERROR = 1,
  WARNING = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface AttributePath {
  steps: AttributePath_Step[];
}

export interface AttributePath_Step {
  /**
   * Set "attribute_name" to represent looking up an attribute
   * in the current object value.
   */
  attributeName?:
    | string
    | undefined;
  /**
   * Set "element_key_*" to represent looking up an element in
   * an indexable collection type.
   */
  elementKeyString?: string | undefined;
  elementKeyInt?: Long | undefined;
}
// ...
export const DynamicValue: MessageFns<DynamicValue> = {
  encode(message: DynamicValue, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.msgpack.length !== 0) {
      writer.uint32(10).bytes(message.msgpack);
    }
    if (message.json.length !== 0) {
      writer.uint32(18).bytes(message.json);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DynamicValue {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDynamicValue();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.msgpack = reader.bytes();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.json = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DynamicValue {
    return {
      msgpack: isSet(object.msgpack) ? bytesFromBase64(object.msgpack) : new Uint8Array(0),
      json: isSet(object.json) ? bytesFromBase64(object.json) : new Uint8Array(0),
    };
  },

  toJSON(message: DynamicValue): unknown {
    const obj: any = {};
    if (message.msgpack.length !== 0) {
      obj.msgpack = base64FromBytes(message.msgpack);
    }
    if (message.json.length !== 0) {
      obj.json = base64FromBytes(message.json);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DynamicValue>, I>>(base?: I): DynamicValue {
    return DynamicValue.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DynamicValue>, I>>(object: I): DynamicValue {
    const message = createBaseDynamicValue();
    message.msgpack = object.msgpack ?? new Uint8Array(0);
    message.json = object.json ?? new Uint8Array(0);
    return message;
  },
};

function createBaseDiagnostic(): Diagnostic {
  return { severity: 0, summary: "", detail: "", attribute: undefined };
}

export const Diagnostic: MessageFns<Diagnostic> = {
  encode(message: Diagnostic, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.severity !== 0) {
      writer.uint32(8).int32(message.severity);
    }
    if (message.summary !== "") {
      writer.uint32(18).string(message.summary);
    }
    if (message.detail !== "") {
      writer.uint32(26).string(message.detail);
    }
    if (message.attribute !== undefined) {
      AttributePath.encode(message.attribute, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Diagnostic {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDiagnostic();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.severity = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.summary = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.detail = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.attribute = AttributePath.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Diagnostic {
    return {
      severity: isSet(object.severity) ? diagnostic_SeverityFromJSON(object.severity) : 0,
      summary: isSet(object.summary) ? globalThis.String(object.summary) : "",
      detail: isSet(object.detail) ? globalThis.String(object.detail) : "",
      attribute: isSet(object.attribute) ? AttributePath.fromJSON(object.attribute) : undefined,
    };
  },

  toJSON(message: Diagnostic): unknown {
    const obj: any = {};
    if (message.severity !== 0) {
      obj.severity = diagnostic_SeverityToJSON(message.severity);
    }
    if (message.summary !== "") {
      obj.summary = message.summary;
    }
    if (message.detail !== "") {
      obj.detail = message.detail;
    }
    if (message.attribute !== undefined) {
      obj.attribute = AttributePath.toJSON(message.attribute);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Diagnostic>, I>>(base?: I): Diagnostic {
    return Diagnostic.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Diagnostic>, I>>(object: I): Diagnostic {
    const message = createBaseDiagnostic();
    message.severity = object.severity ?? 0;
    message.summary = object.summary ?? "";
    message.detail = object.detail ?? "";
    message.attribute = (object.attribute !== undefined && object.attribute !== null)
      ? AttributePath.fromPartial(object.attribute)
      : undefined;
    return message;
  },
};
// ...
export const AttributePath: MessageFns<AttributePath> = {
  encode(message: AttributePath, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.steps) {
      AttributePath_Step.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AttributePath {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAttributePath();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.steps.push(AttributePath_Step.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AttributePath {
    return {
      steps: globalThis.Array.isArray(object?.steps)
        ? object.steps.map((e: any) => AttributePath_Step.fromJSON(e))
        : [],
    };
  },

  toJSON(message: AttributePath): unknown {
    const obj: any = {};
    if (message.steps?.length) {
      obj.steps = message.steps.map((e) => AttributePath_Step.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AttributePath>, I>>(base?: I): AttributePath {
    return AttributePath.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AttributePath>, I>>(object: I): AttributePath {
    const message = createBaseAttributePath();
    message.steps = object.steps?.map((e) => AttributePath_Step.fromPartial(e)) || [];
    return message;
  },
};

function createBaseAttributePath_Step(): AttributePath_Step {
  return { attributeName: undefined, elementKeyString: undefined, elementKeyInt: undefined };
}

export const AttributePath_Step: MessageFns<AttributePath_Step> = {
  encode(message: AttributePath_Step, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.attributeName !== undefined) {
      writer.uint32(10).string(message.attributeName);
    }
    if (message.elementKeyString !== undefined) {
      writer.uint32(18).string(message.elementKeyString);
    }
    if (message.elementKeyInt !== undefined) {
      writer.uint32(24).int64(message.elementKeyInt.toString());
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AttributePath_Step {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAttributePath_Step();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.attributeName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.elementKeyString = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.elementKeyInt = Long.fromString(reader.int64().toString());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AttributePath_Step {
    return {
      attributeName: isSet(object.attributeName)
        ? globalThis.String(object.attributeName)
        : isSet(object.attribute_name)
        ? globalThis.String(object.attribute_name)
        : undefined,
      elementKeyString: isSet(object.elementKeyString)
        ? globalThis.String(object.elementKeyString)
        : isSet(object.element_key_string)
        ? globalThis.String(object.element_key_string)
        : undefined,
      elementKeyInt: isSet(object.elementKeyInt)
        ? Long.fromValue(object.elementKeyInt)
        : isSet(object.element_key_int)
        ? Long.fromValue(object.element_key_int)
        : undefined,
    };
  },

  toJSON(message: AttributePath_Step): unknown {
    const obj: any = {};
    if (message.attributeName !== undefined) {
      obj.attributeName = message.attributeName;
    }
    if (message.elementKeyString !== undefined) {
      obj.elementKeyString = message.elementKeyString;
    }
    if (message.elementKeyInt !== undefined) {
      obj.elementKeyInt = (message.elementKeyInt || Long.ZERO).toString();
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AttributePath_Step>, I>>(base?: I): AttributePath_Step {
    return AttributePath_Step.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AttributePath_Step>, I>>(object: I): AttributePath_Step {
    const message = createBaseAttributePath_Step();
    message.attributeName = object.attributeName ?? undefined;
    message.elementKeyString = object.elementKeyString ?? undefined;
    message.elementKeyInt = (object.elementKeyInt !== undefined && object.elementKeyInt !== null)
      ? Long.fromValue(object.elementKeyInt)
      : undefined;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends Long ? string | number | Long : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}